feat(account): move usecases configuration to usecases folder and add the AccountCodeUseCaseConfig
